+++ src/credential.ts
@@ -47,8 +47,6 @@ export {
   ProvedFromProgram,
 };
 
-export { HashedCredential };
-
 /**
  * A credential is a generic piece of data (the "attributes") along with an owner represented by a public key.
  */
